+++ src/views/DesignEditor/components/ContextMenu/ContextMenu.tsx
@@ -16,38 +16,32 @@ import Ungroup from "~/components/Icons/Ungroup"
 import DownloadIcon from "~/components/Icons/DownloadIcon"
 import DownloadPopup from "../Footer/Graphic/DownloadPopup/DownloadPopup"
 import Elements from "~/components/Icons/Elements"
-import { backgroundLayerType, deviceUploadType } from "~/constants/contants"
+import ContextMenuItem from "./ContextMenuItem"
+import { DuplicateFunc } from "../../utils/tools/DuplicateFunc"
+import { PasteFunc } from "../../utils/tools/PasteFunc"
+import { FrontFunc } from "../../utils/tools/FrontFunc"
+import { BackFunc } from "../../utils/tools/BackFunc"
+import { LockFunc, UnlockFunc } from "../../utils/tools/LockUnlockFunc"
+import { InvisibleFunc, VisibleFunc } from "../../utils/tools/VisibilityFunc"
+import { UngroupFunc } from "../../utils/tools/GroupUngroupFunc"
+import { SetBgFunc } from "../../utils/SetBgFunc"
+import { DeleteFunc } from "../../utils/tools/DeleteFunc"
 
 const ContextMenu = () => {
   const contextMenuRequest = useContextMenuRequest()
   const editor = useEditor()
   const activeObject: any = useActiveObject()
   const { mainImgInfo, setMainImgInfo, setPanelInfo } = useContext(MainImageContext)
-  const frame = useFrame()
 
-  const deleteHandler = () => {
-    if (activeObject?.id === mainImgInfo.id) {
-      // @ts-ignore
-      setPanelInfo((prev) => ({
-        ...prev,
-        uploadSection: true,
-        trySampleImg: true,
-        uploadPreview: false,
-        bgOptions: false,
-        bgRemoverBtnActive: false,
-      }))
-      setMainImgInfo((prev: any) => ({ ...prev, id: "" }))
-    }
-    editor.objects.remove(activeObject?.id)
-  }
-  const handleAsComponentHandler = async () => {
-    if (editor) {
-      const component: any = await editor.scene.exportAsComponent()
-      if (component) {
-        console.log({ component })
-      }
-    }
-  }
+  // const handleAsComponentHandler = async () => {
+  //   if (editor) {
+  //     const component: any = await editor.scene.exportAsComponent()
+  //     if (component) {
+  //       console.log({ component })
+  //     }
+  //   }
+  // }
+
   if (!contextMenuRequest || !contextMenuRequest.target) {
     return <></>
   }
@@ -64,31 +58,18 @@ const ContextMenu = () => {
       >
         <ContextMenuItem
           disabled={true}
-          onClick={() => {
-            editor.objects.copy()
-            editor.cancelContextMenuRequest()
-          }}
+          onClick={() => DuplicateFunc({ editor, activeObject })}
           icon="Duplicate"
           label="Copy"
         >
           <Duplicate size={24} />
         </ContextMenuItem>
-        <ContextMenuItem
-          onClick={() => {
-            editor.objects.paste()
-            editor.cancelContextMenuRequest()
-          }}
-          icon="Paste"
-          label="Paste"
-        >
+        <ContextMenuItem onClick={() => PasteFunc({ editor })} icon="Paste" label="Paste">
           <Paste size={24} />
         </ContextMenuItem>
         <ContextMenuItem
           disabled={true}
-          onClick={() => {
-            deleteHandler()
-            editor.cancelContextMenuRequest()
-          }}
+          onClick={() => DeleteFunc({ editor, activeObject, mainImgInfo, setMainImgInfo })}
           icon="Delete"
           label="Delete"
         >
@@ -107,14 +88,7 @@ const ContextMenu = () => {
           left: `${contextMenuRequest.left}px`,
         }}
       >
-        <ContextMenuItem
-          onClick={() => {
-            editor.objects.copy()
-            editor.cancelContextMenuRequest()
-          }}
-          icon="Duplicate"
-          label="Copy"
-        >
+        <ContextMenuItem onClick={() => DuplicateFunc({ editor, activeObject })} icon="Duplicate" label="Duplicate">
           <Duplicate size={24} />
         </ContextMenuItem>
         {/* <ContextMenuItem
@@ -128,35 +102,16 @@ const ContextMenu = () => {
           <Paste size={24} />
         </ContextMenuItem> */}
         <ContextMenuItem
-          onClick={() => {
-            deleteHandler()
-            editor.cancelContextMenuRequest()
-          }}
+          onClick={() => DeleteFunc({ editor, activeObject, mainImgInfo, setMainImgInfo })}
           icon="Delete"
           label="Delete"
         >
           <Delete size={24} />
         </ContextMenuItem>
-        <ContextMenuItem
-          onClick={() => {
-            editor.objects.bringForward()
-            editor.objects.update({ name: activeObject.name })
-            editor.cancelContextMenuRequest()
-          }}
-          icon="Forward"
-          label="Bring forward"
-        >
+        <ContextMenuItem onClick={() => FrontFunc({ editor, activeObject })} icon="Forward" label="Bring forward">
           <ArrowUp />
         </ContextMenuItem>
-        <ContextMenuItem
-          onClick={() => {
-            editor.objects.sendBackwards()
-            editor.objects.update({ name: activeObject.name })
-            editor.cancelContextMenuRequest()
-          }}
-          icon="Backward"
-          label="Send backward"
-        >
+        <ContextMenuItem onClick={() => BackFunc({ editor, activeObject })} icon="Backward" label="Send backward">
           <SendBack />
         </ContextMenuItem>
         {/* <ContextMenuItem
@@ -172,96 +127,31 @@ const ContextMenu = () => {
         <div style={{ margin: "0.5rem 0" }} />
 
         {!contextMenuRequest.target.locked ? (
-          <ContextMenuItem
-            onClick={() => {
-              editor.objects.lock()
-              editor.cancelContextMenuRequest()
-            }}
-            icon="Locked"
-            label="Lock"
-          >
+          <ContextMenuItem onClick={() => LockFunc({ editor })} icon="Locked" label="Lock">
             <Unlocked size={22} />
           </ContextMenuItem>
         ) : (
-          <ContextMenuItem
-            onClick={() => {
-              editor.objects.unlock()
-              editor.cancelContextMenuRequest()
-            }}
-            icon="Unlocked"
-            label="Unlock"
-          >
+          <ContextMenuItem onClick={() => UnlockFunc({ editor })} icon="Unlocked" label="Unlock">
             <Unlocked size={22} />
           </ContextMenuItem>
         )}
 
         {activeObject?.visible === true ? (
-          <ContextMenuItem
-            onClick={() => {
-              editor.objects.update({ visible: false })
-              editor.cancelContextMenuRequest()
-            }}
-            icon="eye"
-            label="Visibility"
-          >
+          <ContextMenuItem onClick={() => InvisibleFunc({ editor })} icon="eye" label="Visibility">
             <Eye size={24} />
           </ContextMenuItem>
         ) : (
-          <ContextMenuItem
-            onClick={() => {
-              editor.objects.update({ visible: true })
-              editor.cancelContextMenuRequest()
-            }}
-            icon="eye"
-            label="Visibility"
-          >
+          <ContextMenuItem onClick={() => VisibleFunc({ editor })} icon="eye" label="Visibility">
             <EyeCrossed size={24} />
           </ContextMenuItem>
         )}
         {activeObject?.type === "group" && (
-          <ContextMenuItem
-            onClick={() => {
-              editor.objects.ungroup()
-              editor.cancelContextMenuRequest()
-            }}
-            icon="layers"
-            label="Ungroup"
-          >
+          <ContextMenuItem onClick={() => UngroupFunc({ editor })} icon="layers" label="Ungroup">
             <Ungroup size={26} />
           </ContextMenuItem>
         )}
         {activeObject?.type === "StaticImage" && (
-          <ContextMenuItem
-            onClick={() => {
-              // handleAsComponentHandler()
-              editor.frame.resize({ width: frame.width, height: frame.height })
-              const bgObject = editor?.frame?.background?.canvas?._objects.filter(
-                (el: any) => el.type === "BackgroundImage"
-              )[0]
-
-              const deviceUploadImg = editor?.frame?.background?.canvas?._objects.filter(
-                (el: any) => el.metadata?.type === deviceUploadType || el.metadata?.type === backgroundLayerType
-              )[0]
-
-              if (bgObject) {
-                editor.objects.removeById(bgObject.id)
-              } else if (deviceUploadImg) {
-                editor.objects.removeById(deviceUploadImg.id)
-              }
-              editor.objects.unsetBackgroundImage()
-              const options = {
-                type: "BackgroundImage",
-                metadata: { generationDate: new Date().getTime(), type: deviceUploadType },
-              }
-
-              editor.objects.setAsBackgroundImage()
-              editor.objects.update({ ...options })
-
-              editor.cancelContextMenuRequest()
-            }}
-            icon="Images"
-            label="Set as background image"
-          >
+          <ContextMenuItem onClick={() => SetBgFunc({ editor })} icon="Images" label="Set as background image">
             <Elements size={24} />
           </ContextMenuItem>
         )}
@@ -281,36 +171,6 @@ const ContextMenu = () => {
         <div className={clsx(classes.chevronTopIcon)}>
           <Icons.SliderBtn size={20} width="10" />
         </div>
-      </div>
-    </div>
-  )
-}
-
-const ContextMenuItem = ({
-  label,
-  icon,
-  onClick,
-  children,
-  disabled = false,
-}: {
-  icon: string
-  label: string
-  onClick: () => void
-  disabled?: boolean
-  children: React.ReactNode
-}) => {
-  return (
-    <div>
-      <div onClick={onClick} className={clsx(classes.eachMenu, disabled && classes.disabledMenu)}>
-        <div style={{ width: "25px" }} className={"d-flex justify-content-center mr-2"}>
-          {children}
-        </div>{" "}
-        {label}
-        {icon === "download" && (
-          <div className={classes.rightIcon}>
-            <Icons.SliderIcon size={15} />
-          </div>
-        )}
       </div>
     </div>
   )
