+++ src/views/DesignEditor/utils/tools/VisibilityFunc.tsx
@@ -1,11 +1,11 @@
 const InvisibleFunc = ({ editor, setState }: any) => {
   editor.objects.update({ visible: false })
-  setState({ visible: false })
+  if (setState) setState({ visible: false })
   editor.cancelContextMenuRequest()
 }
 const VisibleFunc = ({ editor, setState }: any) => {
   editor.objects.update({ visible: true })
-  setState({ visible: true })
+  if (setState) setState({ visible: true })
   editor.cancelContextMenuRequest()
 }
